% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPplot1.R
\name{DPplot1}
\alias{DPplot1}
\title{Dynamic prediction plot for one individual}
\usage{
DPplot1(
  object,
  s = s,
  id_new = id_new,
  mi = mi,
  by = 0.1,
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = floor(n.iter/2),
  dataLong,
  dataSurv
)
}
\arguments{
\item{object}{an object inheriting from class ZIJMCV}

\item{s}{the landmark time for prediction}

\item{mi}{the number of multiple imputation for Monte-Carlo approximation; default is 10.}

\item{by}{number: increment of the sequence of DP.}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}
}
\value{
\itemize{
\item A plot for DP
}
}
\description{
Dynamic prediction plot for ZIJMCV
}
\details{
Estimate DP for joint modeling based on ZIJMCV
}
\examples{
library(survival)
rm(list=ls())

\donttest{

  set.seed(2)
  INDTRAIN <- sample(surv_data_nb$id, 0.7 * (dim(surv_data_nb)[1]))
  INDVALID <- surv_data_nb$id[-INDTRAIN]
  dataLong_t <- subset(
    long_data_nb,
    long_data_nb$id \%in\% INDTRAIN
  )
  dataSurv_t <- subset(
    surv_data_nb,
    surv_data_nb$id \%in\% INDTRAIN
  )
  names(dataSurv_t)

  dataLong_v <- subset(
    long_data_nb,
    long_data_nb$id \%in\% INDVALID
  )
  dataSurv_v <- subset(
    surv_data_nb,
    surv_data_nb$id \%in\% INDVALID
  )


  Z2 <- ZIJMCV(
    FixedY = Y1 ~ obstime + x1 + x2, RandomY = ~obstime, GroupY = ~id,
    FixedZ = ~ obstime + x1, RandomZ = ~obstime, GroupZ = ~id,
    formSurv = Surv(survtime, death) ~ w1 + w2,
    dataLong = dataLong_t, dataSurv = dataSurv_t,
    obstime = "obstime", id = "id", n.chains = 2,
    n.iter = 200, n.burnin = 100, n.thin = 1, K = 15, family = "NB"
  )


  DD <- DP_CV(
    object = Z2, s = 0.5, t = 0.5, n.chains = 1, n.iter = 2000, n.burnin = 1000,
    n.thin = 1, dataLong = dataLong_v, dataSurv = dataSurv_v
  )


  DPplot1(Z2, s = 0.4, id_new=498, by=0.2, n.chains = 1, n.iter = 200, n.burnin =100,
          dataLong = dataLong_v, dataSurv = dataSurv_v)


}



\donttest{
  rm(list = ls())
 INDTRAIN <- sample(surv_data_n$id, 0.7 * (dim(surv_data_n)[1]))
 INDVALID <- surv_data_n$id[-INDTRAIN]
 dataLong_t <- subset(
   long_data_n,
   long_data_n$id \%in\% INDTRAIN
 )
 dataSurv_t <- subset(
   surv_data_n,
   surv_data_n$id \%in\% INDTRAIN
 )
 names(dataSurv_t)

 dataLong_v <- subset(
   long_data_n,
   long_data_n$id \%in\% INDVALID
 )
 dataSurv_v <- subset(
   surv_data_n,
   surv_data_n$id \%in\% INDVALID
 )

 Z1 <- ZIJMCV(
   FixedY = Y1 ~ obstime + x1 + x2, RandomY = ~obstime, GroupY = ~id,
   FixedZ = ~ obstime + x1, RandomZ = ~obstime, GroupZ = ~id,
   formSurv = survival::Surv(survtime, death) ~ w1,
   dataLong = dataLong_t, dataSurv = dataSurv_t,
   obstime = "obstime", id = "id",
   n.chains = 2,
   n.iter = 2000, n.burnin = 1000, n.thin = 1, K = 15, family = "Gaussian"
 )


 DD <- DP_CV(
   object = Z1, s = 0.5, t = 0.5, n.chains = 1, n.iter = 1000, n.burnin = 500,
   n.thin = 1, dataLong = dataLong_v, dataSurv = dataSurv_v
 )



  }
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}
}
