% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_data_e.R
\docType{data}
\name{long_data_e}
\alias{long_data_e}
\title{Simulated longitudinal data}
\format{
A data frame which contains id, obstime, x1, x2, and Y1.
\describe{
  \item{id}{patients identifier}
  \item{obstime}{observed times for longitudinal measurements}
  \item{Y1}{the longitudinal response}
  \item{x1}{a continuous covariate}
  \item{x2}{a Binary covariate}
}
}
\usage{
long_data_e
}
\description{
Simulated data were generated, where "Y1" follows a hurdle exponential distribution, and "obstime" represents the observed time with two covariates, x1 and x2.
}
\seealso{
\code{\link{UHJM}}
}
\keyword{datasets}
