% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_data_nb.R
\docType{data}
\name{long_data_nb}
\alias{long_data_nb}
\title{Simulated longitudinal data}
\format{
A data frame which contains id, obstime, x1, x2, and Y1.
\describe{
  \item{id}{patients identifier}
  \item{obstime}{observed times for longitudinal measurements}
  \item{Y1}{the longitudinal response}
  \item{x1}{a continuous covariate}
  \item{x2}{a Binary covariate}
}
}
\usage{
long_data_nb
}
\description{
Simulated data were generated in which "Y1" follows a hurdle negative binomial distribution, "obstime" represents the observed time with two covariates x1 and x2.
}
\seealso{
\code{\link{UHJM}}
}
\keyword{datasets}
