% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_data_p.R
\docType{data}
\name{surv_data_p}
\alias{surv_data_p}
\title{Simulated survival data}
\format{
A data frame which contains id, survtime, death, w1.
\describe{
  \item{id}{patients identifier}
  \item{survtime}{survival time (the response variable)}
  \item{death}{censoring indicator, 1=observed, 0=censored}
  \item{w1}{a continuous covariate}
  \item{w2}{a binary covariate}
}
}
\usage{
surv_data_p
}
\description{
Simulated survival data were generated in which "survtime" follows a survival time with a censoring indicator called "death" and a continuous covariate, "w1".
}
\seealso{
\code{\link{UHJM}}
}
\keyword{datasets}
