% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHJM.R
\docType{package}
\name{UHJM}
\alias{UHJM-package}
\alias{UHJM}
\title{UHJM Package}
\description{
Performing a Gibbs sampler for hurdle joint models involves estimating joint models for zero-inflated longitudinal measurements and time-to-event data. This versatile package accommodates various distributional assumptions, including Gaussian, Gamma, inverse Gaussian, Weibull, exponential, beta, Poisson, negative binomial, logarithmic, Bell, generalized Poisson, and binomial. As described in Ganjali et al. (2024), these models are implemented through two key functions: (1) "ZIJM" facilitates joint modeling with a proportional hazard sub-model and a piecewise constant baseline hazard, considering associations based on the current values, and (2) "ZISRE" enables joint modeling with a Weibull sub-model by incorporating a shared random effects model.
}
\references{
Ganjali, M. & Baghfalaki, T. & Balakrishnan, N. (2024). Joint Modeling of Zero-Inflated Longitudinal Measurements and Time-to-Event Outcome with Applications in Dynamic Prediction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tbaghfalaki/UHJM}
}

}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}, Mojtaba Ganjali \email{m-ganjali@sbu.ac.ir}
}
\keyword{UHJM}
