% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZISRE.R
\name{ZISRE}
\alias{ZISRE}
\title{Zero-inflation joint modeling}
\usage{
ZISRE(
  FixedY,
  RandomY,
  GroupY,
  FixedZ,
  RandomZ,
  GroupZ,
  formSurv,
  dataLong,
  dataSurv,
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = n.burnin,
  n.thin = n.thin,
  family = "Poisson"
)
}
\arguments{
\item{FixedY}{formula for fixed part of longitudinal count model}

\item{RandomY}{formula for random part of longitudinal count model}

\item{GroupY}{formula specifying the cluster variable for Y (e.g. = ~ subject)}

\item{FixedZ}{formula for fixed part of longitudinal probability model}

\item{RandomZ}{formula for random part of longitudinal probability model}

\item{GroupZ}{formula specifying the cluster variable for Z (e.g. = ~ subject)}

\item{formSurv}{formula for survival model}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{n.thin}{integer specifying the thinning of the chains; default is 1.}

\item{family}{Family objects provide a convenient way to specify the details of the models. They cover various distributions like "Gaussian", "Exponential", "Weibull", "Gamma", "Beta", "inverse.gaussian", "Poisson", "NB", "Logarithmic", "Bell", "GP", and "Binomial". Specifically, "NB" and "GP" are tailored for hurdle negative binomial and hurdle generalized Poisson joint models, respectively, while the others are utilized for the corresponding models based on their names.}
}
\value{
\itemize{
\item mu.vect list of posterior mean for each parameter
\item sd.vect list of standard error for each parameter
\item 2.5\% list of posterior mode for each parameter
\item 25\% list of posterior median for each parameter
\item 50\% list of posterior median for each parameter
\item 75\% list of posterior median for each parameter
\item 97.5\% list of posterior median for each parameter
\item Rhat Gelman and Rubin diagnostic for all parameter
\item betaL1 the regression coefficients of rate of the hurdle model
\item betaL2 the regression coefficients of probability of the hurdle model
\item betaS the regression coefficients of the survival model
\item Sigma the variance of random effects
\item gamma the association parameters
}
}
\description{
Fits zero-inflated hurdle shared random effects models under Gaussian, Gamma, inverse Gaussian, Weibull, exponential, beta, Poisson, negative binomial, logarithmic, Bell, generalized Poisson, and binomial.
}
\details{
Function using 'JAGS' software to estimate the hurdle joint modeling
}
\examples{
library(survival)


\donttest{
  data(long_data_nb)
  data(surv_data_nb)
Z1 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_nb, dataSurv = surv_data_nb,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Poisson"
)


Z2 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_nb, dataSurv = surv_data_nb,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "NB"
)

Z3 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_nb, dataSurv = surv_data_nb,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "GP"
)


Z4 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_nb, dataSurv = surv_data_nb,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Bell"
)

Z5 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_nb, dataSurv = surv_data_nb,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Logarithmic"
)


Z6 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_nb, dataSurv = surv_data_nb,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "binomial"
)
}





\donttest{

  data(long_data_n)
  data(surv_data_n)
Z1 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_n, dataSurv = surv_data_n,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Gaussian"
)

data(long_data_b)
data(surv_data_b)
Z2 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_b, dataSurv = surv_data_b,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Beta"
)

data(long_data_e)
data(surv_data_e)
Z3 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_e, dataSurv = surv_data_e,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Exponential"
)


Z4 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_e, dataSurv = surv_data_e,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Weibull"
)

Z5 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_e, dataSurv = surv_data_e,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Gamma"
)


Z6 <- ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long_data_e, dataSurv = surv_data_e,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "inverse.gaussian"
)
}
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}, Mojtaba Ganjali \email{m-ganjali@sbu.ac.ir}
}
